pyproject.toml
[tool.poetry]
name = "logic_and_flow_control"
version = "0.1.0"
description = ""
authors = ["peterdougstuart <peter.stuart@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
pygame = "^2.6.1"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
